Update code based on suggestions from `cargo clippy --fix`
Added default() associated function for APIstuff struct and updated main in octo_web.rs
and lib.rs to use this instead of the new() associated function.

Updated return type of construct_uri to no longer be a of type
`Result<_, ()>` is now of type `String`. This means the .unwrap() method
no longer needs to be called on construct_uri.

Remove leftover TODO comment in URIstuff's new associated function